feat(loader): switch from Dna to Triangle spinner
Replace the DNA loader animation with a Triangle spinner in the ListaPostagens component. This change updates the user interface to use a 'purple' Triangle loading animation instead of the previous Dna loader, providing a refreshed visual experience during data fetching states. The `color` attribute is set to "purple", the `ariaLabel` is updated to "triangle-loading", and the `wrapperClass` is revised to align with the new Triangle spinner usage.

The size of the width for the Triangle spinner seems to have been mistakenly changed from "200" to "2000". This might need to be reviewed if it was not an intentional adjustment.